requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.2.1
psycopg2-binary>=2.9.7

# Para desarrollo
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2